+++ shared/homebrewery/renderWarnings/renderWarnings.jsx
@@ -3,7 +3,7 @@ const React = require('react');
 const _     = require('lodash');
 const cx    = require('classnames');
 
-const Warnings = React.createClass({
+const RenderWarnings = React.createClass({
 	getInitialState: function() {
 		return {
 			warnings: {}
@@ -18,21 +18,27 @@ const Warnings = React.createClass({
 	warnings : {
 		chrome : function(){
 			const isChrome = /Chrome/.test(navigator.userAgent) && /Google Inc/.test(navigator.vendor);
-			if(isChrome){
-				return 'OPtimized for Chrome';
+			if(!isChrome){
+				return <li key='chrome'>
+					<em>Built for Chrome </em> <br />
+					Other browsers do not support key features this site uses.
+				</li>;
 			}
 		},
 		zoom : function(){
 			if(window.devicePixelRatio !== 1){
-				return 'You are zoomed';
+				return <li key='zoom'>
+					<em>Your browser is zoomed. </em> <br />
+					This can cause content to jump columns.
+				</li>;
 			}
 		}
 	},
 	checkWarnings : function(){
 		this.setState({
 			warnings : _.reduce(this.warnings, (r, fn, type) => {
-				const text = fn();
-				if(text) r[type] = text;
+				const element = fn();
+				if(element) r[type] = element;
 				return r;
 			}, {})
 		})
@@ -40,17 +46,13 @@ const Warnings = React.createClass({
 	render: function(){
 		if(_.isEmpty(this.state.warnings)) return null;
 
-		const errors = _.map(this.state.warnings, (text, idx) => {
-			return <li key={idx}>{text}</li>
-		});
-
-		return <div className='warnings'>
+		return <div className='renderWarnings'>
 			<i className='fa fa-exclamation-triangle' />
-			<h3>Rendering Warnings</h3>
+			<h3>Render Warnings</h3>
 			<small>If this homebrew is rendering badly if might be because of the following:</small>
-			<ul>{errors}</ul>
+			<ul>{_.values(this.state.warnings)}</ul>
 		</div>
 	}
 });
 
-module.exports = Warnings;
+module.exports = RenderWarnings;